requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
cloudinary==1.26.0
Django==3.2.4
django-admin-colors==0.2.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.12.0
django-embed-video==1.4.0
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
